+++ resources/release-notes.md
@@ -1,6 +1,6 @@
 # Software
 
-- 
+- Update system test `tIMAPAnalysis` to cover range changing
 
 # Project